Refactor TaskMap styling and add localStorage functionality for checkbox state
