+++ src/TaskMap/hooks.tsx
@@ -7,61 +7,12 @@ import {
   useState,
 } from "react";
 import dagre from "dagre";
-import ReactFlow, { Controls, Edge, MiniMap, Node, Position } from "reactflow";
-import styled from "styled-components";
+import { Edge, Node, Position } from "reactflow";
 import { CustomNode } from "./CustomNodes";
 import { CustomEdge } from "./CustomEdge";
 import { LanguageDictContext } from "@/App";
 import { GetTestDummy } from "./GetTestDummy";
 
-const darkTheme = {
-  bg: "#000",
-  primary: "#ff0072",
-
-  nodeBg: "#343435",
-  nodeColor: "#f9f9f9",
-  nodeBorder: "#888",
-
-  minimapMaskBg: "#343435",
-
-  controlsBg: "#555",
-  controlsBgHover: "#676768",
-  controlsColor: "#dddddd",
-  controlsBorder: "#676768",
-};
-
-const ReactFlowStyled = styled(ReactFlow)`
-  background-color: ${darkTheme.bg};
-`;
-
-const MiniMapStyled = styled(MiniMap)`
-  background-color: ${darkTheme.bg};
-
-  .react-flow__minimap-mask {
-    fill: ${darkTheme.minimapMaskBg};
-  }
-
-  .react-flow__minimap-node {
-    fill: ${darkTheme.nodeBg};
-    stroke: none;
-  }
-`;
-
-const ControlsStyled = styled(Controls)`
-  button {
-    background-color: ${darkTheme.controlsBg};
-    color: ${darkTheme.controlsColor};
-    border-bottom: 1px solid ${darkTheme.controlsBorder};
-
-    &:hover {
-      background-color: ${darkTheme.controlsBgHover};
-    }
-
-    path {
-      fill: currentColor;
-    }
-  }
-`;
 type dataType = {
   nodes: Node[];
   edges: Edge[];
@@ -234,10 +185,6 @@ export const useHooks = () => {
     isLoading,
     showKappaRequired,
     handleCheckboxChange,
-    darkTheme,
-    ReactFlowStyled,
-    MiniMapStyled,
-    ControlsStyled,
     langDict,
   };
 };
